+++ pkg/commands/service/logging/openstack/openstack_test.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"testing"
 
-	"github.com/fastly/go-fastly/v16/fastly"
+	"github.com/fastly/go-fastly/v17/fastly"
 
 	"github.com/fastly/cli/pkg/argparser"
 	"github.com/fastly/cli/pkg/commands/service/logging/openstack"
@@ -135,7 +135,38 @@ func TestUpdateOpenstackInput(t *testing.T) {
 				ResponseCondition: fastly.ToPointer("new8"),
 				MessageType:       fastly.ToPointer("new9"),
 				TimestampFormat:   fastly.ToPointer("new10"),
-				Placement:         fastly.ToPointer("new11"),
+				Placement:         fastly.NewNullable("new11"),
+				PublicKey:         fastly.ToPointer("new12"),
+				CompressionCodec:  fastly.ToPointer("new13"),
+				ProcessingRegion:  fastly.ToPointer("eu"),
+			},
+		},
+		{
+			name: "reset placement to null",
+			cmd:  updateCommandPlacementReset(),
+			api: mock.API{
+				GetVersionFn:   testutil.GetVersion,
+				CloneVersionFn: testutil.CloneVersionResult(4),
+				GetOpenstackFn: getOpenstackOK,
+			},
+			want: &fastly.UpdateOpenstackInput{
+				ServiceID:         "123",
+				ServiceVersion:    4,
+				Name:              "log",
+				NewName:           fastly.ToPointer("new1"),
+				BucketName:        fastly.ToPointer("new2"),
+				User:              fastly.ToPointer("new3"),
+				AccessKey:         fastly.ToPointer("new4"),
+				URL:               fastly.ToPointer("new5"),
+				Path:              fastly.ToPointer("new6"),
+				Period:            fastly.ToPointer(3601),
+				GzipLevel:         fastly.ToPointer(0),
+				Format:            fastly.ToPointer("new7"),
+				FormatVersion:     fastly.ToPointer(3),
+				ResponseCondition: fastly.ToPointer("new8"),
+				MessageType:       fastly.ToPointer("new9"),
+				TimestampFormat:   fastly.ToPointer("new10"),
+				Placement:         fastly.NullValue[string](),
 				PublicKey:         fastly.ToPointer("new12"),
 				CompressionCodec:  fastly.ToPointer("new13"),
 				ProcessingRegion:  fastly.ToPointer("eu"),
@@ -382,6 +413,12 @@ func updateCommandAll() *openstack.UpdateCommand {
 		CompressionCodec:  argparser.OptionalString{Optional: argparser.Optional{WasSet: true}, Value: "new13"},
 		ProcessingRegion:  argparser.OptionalString{Optional: argparser.Optional{WasSet: true}, Value: "eu"},
 	}
+}
+
+func updateCommandPlacementReset() *openstack.UpdateCommand {
+	c := updateCommandAll()
+	c.Placement = argparser.OptionalString{Optional: argparser.Optional{WasSet: true}, Value: ""}
+	return c
 }
 
 func updateCommandMissingServiceID() *openstack.UpdateCommand {