+++ pkg/text/healthcheck.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/segmentio/textio"
 
-	"github.com/fastly/go-fastly/v16/fastly"
+	"github.com/fastly/go-fastly/v17/fastly"
 )
 
 // PrintHealthCheck pretty prints a fastly.HealthCheck structure in verbose